Change object ids to use uint8_t to avoid name collisions

+++ Novella/Input/ActionCommand.hpp
@@ -1,10 +1,11 @@
 #pragma once
 #include <string>
 #include <nlohmann/json.hpp>
+#include <cstdint>
 
 namespace Novella::Input{
 
-    enum class Target : unsigned int;
+    enum class Target : uint64_t;
 
     struct ActionCommand{
